cleanup: remove deeplink building to external hook
